pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openapi2locust"
version = "0.1.0"
description = "Tool for generating locust scripts based on openapi specs"
authors = [{name = "OpenAPI2Locust Team"}]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Testing :: Traffic Generation",
]
dependencies = [
    "locust>=2.0.0",
    "openapi-spec-validator>=0.5.0",
    "pyyaml>=6.0",
    "jinja2>=3.0.0",
    "faker>=18.0.0",
    "click>=8.0.0",
    "jsonschema>=4.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "types-PyYAML",
    "types-requests",
]

[project.scripts]
openapi2locust = "openapi2locust.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/openapi2locust"
Repository = "https://github.com/yourusername/openapi2locust"
Issues = "https://github.com/yourusername/openapi2locust/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true